requirements.txt
numpy 
ipykernel 
sklearn.preprocessing